automator/requirements.txt
dnspython
requests
Django==4.2.11
django-ninja